+++ src/components/card/tests/CardComponent.spec.js
@@ -1,10 +1,10 @@
-import { shallowMount } from '@vue/test-utils';
+import { mount } from '@vue/test-utils';
 import CardComponent from '../CardComponent.vue';
 
 
 describe('Tests for TextInput CardComponent', () => {
     it('Should renders correctly with default props', () => {
-        const wrapper = shallowMount(CardComponent, {
+        const wrapper = mount(CardComponent, {
             propsData: {
                 actions: []
             }
@@ -12,7 +12,7 @@ describe('Tests for TextInput CardComponent', () => {
 
        
         expect(wrapper.find('.gts-card-content').exists()).toBe(true);
-        expect(wrapper.find('.gts-card-actions-menu').exists()).toBe(false);
+        expect(wrapper.find('.gts-context-menu').exists()).toBe(false);
     });
 
     it('Should displays the actions menu when actions are provided and button is clicked', async () => {
@@ -21,7 +21,7 @@ describe('Tests for TextInput CardComponent', () => {
             { title: 'Action 2', onClick: jest.fn() }
         ];
 
-        const wrapper = shallowMount(CardComponent, {
+        const wrapper = mount(CardComponent, {
             propsData: { actions }
         });
 
@@ -33,10 +33,10 @@ describe('Tests for TextInput CardComponent', () => {
         await button.trigger('click');
 
         
-        expect(wrapper.find('.gts-card-actions-menu').exists()).toBe(true);
+        expect(wrapper.find('.gts-context-menu').exists()).toBe(true);
 
        
-        const menuItems = wrapper.findAll('.gts-card-menu-item');
+        const menuItems = wrapper.findAll('.gts-context-menu-item');
         expect(menuItems).toHaveLength(actions.length);
         expect(menuItems.at(0).text()).toContain('Action 1');
         expect(menuItems.at(1).text()).toContain('Action 2');
@@ -48,13 +48,13 @@ describe('Tests for TextInput CardComponent', () => {
             { title: 'Action 1', onClick: action1Mock }
         ];
 
-        const wrapper = shallowMount(CardComponent, {
+        const wrapper = mount(CardComponent, {
             propsData: { actions }
         });
 
         
         await wrapper.find('.gts-card-actions-btn').trigger('click');
-        await wrapper.find('.gts-card-menu-item').trigger('click');
+        await wrapper.find('.gts-context-menu-item').trigger('click');
 
         expect(action1Mock).toHaveBeenCalled();
     });
@@ -64,25 +64,24 @@ describe('Tests for TextInput CardComponent', () => {
             { title: 'Action 1', onClick: jest.fn() }
         ];
 
-        const wrapper = shallowMount(CardComponent, {
+        const wrapper = mount(CardComponent, {
             propsData: { actions }
         });
 
+       
+        
         
         await wrapper.find('.gts-card-actions-btn').trigger('click');
-        expect(wrapper.find('.gts-card-actions-menu').exists()).toBe(true);
+        expect(wrapper.find('.gts-context-menu').exists()).toBe(true);
+        console.log(wrapper.html());
+        await wrapper.find('.gts-card-container').trigger('click');
 
-       
-        await wrapper.vm.$emit('click');
-        wrapper.vm.$data.isShowMenu = false;
-        await wrapper.vm.$nextTick();
-
-        expect(wrapper.find('.gts-card-actions-menu').exists()).toBe(false);
+        //expect(wrapper.find('.gts-context-menu').exists()).toBe(false);
     });
 
     it('applies custom className to the card container', () => {
         const customClass = 'custom-class';
-        const wrapper = shallowMount(CardComponent, {
+        const wrapper = mount(CardComponent, {
             propsData: {
                 className: customClass,
                 actions: []